changed javax.servlet to jakarta.servlet

+++ axis-rt-core/src/main/java/org/apache/axis/types/UnsignedInt.java
@@ -107,7 +107,7 @@ public class UnsignedInt extends java.lang.Number implements java.lang.Comparabl
     // implement java.lang.comparable interface
     public int compareTo(Object obj) {
       if (lValue != null)
-        return lValue.compareTo(obj);
+        return lValue.compareTo((Long) obj);
       else
         if (equals(obj) == true)
             return 0;  // null == null
