+++ axis-rt-core/src/main/java/org/apache/axis/types/UnsignedLong.java
@@ -101,7 +101,7 @@ public class UnsignedLong extends java.lang.Number
     // implement java.lang.comparable interface
     public int compareTo(Object obj) {
         if (lValue != null)
-            return lValue.compareTo(obj);
+            return lValue.compareTo((BigInteger) obj);
         else if (equals(obj) == true)
             return 0;  // null == null
         else
